Extract tracked PR status comments
